requirements.txt
clearcustomtkinter==5.2.2
pyperclip==1.8.2